% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_constraints.R
\name{get_fs_ct}
\alias{get_fs_ct}
\title{Find central tendency of feasible set}
\usage{
get_fs_ct(s, n, nsamples)
}
\arguments{
\item{s}{how many species}

\item{n}{how many individuals}

\item{nsamples}{how many samples to use}
}
\value{
vector central tendency of feasible set
}
\description{
Use \code{feasiblesads} to generate samples from the feasible set and find the central tendency
}
