% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_constraints.R
\name{get_mete_prediction}
\alias{get_mete_prediction}
\title{Get METE prediction as RAD}
\usage{
get_mete_prediction(s, n)
}
\arguments{
\item{s}{how many species}

\item{n}{how many individuals}
}
\value{
vector METE prediction as an RAD
}
\description{
Use \code{meteR} to get the METE prediction for the SAD as a vector
}
