% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_emp_constraint.R
\name{get_stat_list}
\alias{get_stat_list}
\title{Get test stats for empirical and sampled RADs}
\usage{
get_stat_list(empirical_rad, sampled_rads, stat = "r2",
  constraint = NULL)
}
\arguments{
\item{empirical_rad}{vector RAD for empirical community}

\item{sampled_rads}{matrix RAD for sampled communities}

\item{stat}{character "r2", "evar", "simpson", "rad_skew", "kl_div"}

\item{constraint}{if stat = r2 or kl_div, vector comparison RAD (central tendency of feasible set, or METE prediction)}
}
\value{
list of empirical test stat, sampled test stats
}
\description{
Wrapper for stats functions in conditionalsads.
}
