% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_emp_constraint.R
\name{kl_div}
\alias{kl_div}
\title{Calculate a single K-L divergence}
\usage{
kl_div(focal_rad, compare_rad)
}
\arguments{
\item{focal_rad}{ascending RAD ("empirical")}

\item{compare_rad}{ascending RAD from samples}
}
\value{
K-L div
}
\description{
Calculate a single K-L divergence
}
