% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_portal_plants.R
\name{load_portal_plants}
\alias{load_portal_plants}
\title{Download scripts and load data from Supp et al 2013}
\usage{
load_portal_plants(download = TRUE)
}
\arguments{
\item{download}{TRUE/FALSE, do the files need to be downloaded}
}
\value{
list of data frames for winter and summer plant communities
}
\description{
Download files from https://github.com/weecology/portal-experimental-macroeco and load Portal plant data into workspace. Stores files in \code{analysis/supp-2013-files}.
}
