% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_toy_data.R
\name{make_toy_data}
\alias{make_toy_data}
\title{Simulate toy data}
\usage{
make_toy_data(constraint = "FS", nsamples = 50, ctrl_s = 10,
  ctrl_n = 1000)
}
\arguments{
\item{nsamples}{how many samples for each treatment/control pair?}

\item{param}{constraint "METE" or "FS", whether to draw from METE or feasible set for constraint}
}
\value{
list of 4 long data frames with the above properties
}
\description{
Simulates toy data:
Toy data should have:
Comparisons (difference btwn treatment and control):
h1: Same S and N and both are drawn from the constraint
h2: S and N change but both are drawn from the constraint
h3: Same S and N but treatment is different from constraint
h4: S and N change and treatment is different from constraint
Draw from FS and see if METE gives same result
And vice versa
}
