% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_portal_plants.R
\name{process_portal_plants}
\alias{process_portal_plants}
\title{Process Portal plant data}
\usage{
process_portal_plants(portal_plants)
}
\arguments{
\item{portal_plants}{output of \code{load_portal_plants}}
}
\value{
matrix ready for use
}
\description{
Process Portal plant data
}
