% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_emp_constraint.R
\name{r2}
\alias{r2}
\title{Calculate a single r2}
\usage{
r2(focal_rad, compare_rad)
}
\arguments{
\item{focal_rad}{ascending RAD ("empirical")}

\item{compare_rad}{ascending RAD from samples}
}
\value{
r2
}
\description{
Calculate a single r2
}
