% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_emp_constraint.R
\name{rad_poilog_cs}
\alias{rad_poilog_cs}
\title{Poilog parameters (From Supp et al 2013)}
\usage{
rad_poilog_cs(x_vec)
}
\arguments{
\item{x_vec}{rad}
}
\value{
exp(mu) and sig
}
\description{
Poilog parameters (From Supp et al 2013)
}
