% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_emp_constraint.R
\name{rad_skew}
\alias{rad_skew}
\title{Calculate a single skewness}
\usage{
rad_skew(focal_rad)
}
\arguments{
\item{focal_rad}{RAD}
}
\value{
focal_skewness
}
\description{
Calculate a single skewness
}
