% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_constraints.R
\name{sample_feasibleset}
\alias{sample_feasibleset}
\title{Draw from feasible set}
\usage{
sample_feasibleset(s, n, nsamples)
}
\arguments{
\item{s}{how many species}

\item{n}{how many individuals}

\item{nsamples}{how many samples to draw}
}
\value{
matrix of samples. rows are samples, columns are species
}
\description{
Use \code{feasiblesads} to generate samples from the feasible set
}
