% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_constraints.R
\name{sample_plant_constraints}
\alias{sample_plant_constraints}
\title{Sample plant constraints}
\usage{
sample_plant_constraints(plant_abund_row, nsamples, fs = TRUE,
  mete = TRUE)
}
\arguments{
\item{plant_abund_row}{an RAD, NAs ok}

\item{nsamples}{how many samples}

\item{fs}{sample from fs?}

\item{mete}{sample from mete?}
}
\value{
list of samples
}
\description{
Sample plant constraints
}
