% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_emp_constraint.R
\name{simpson}
\alias{simpson}
\title{Calculate Simpson evenness}
\usage{
simpson(focal_rad)
}
\arguments{
\item{focal_rad}{ascending RAD}
}
\value{
Simpson evenness (1-D) for the RAD
}
\description{
Calculate Simpson evenness
}
