% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_emp_constraint.R
\name{test_quantile}
\alias{test_quantile}
\title{Calculate the quantile of a test stat for a focal RAD vs. a sample of RADs}
\usage{
test_quantile(stats_list)
}
\arguments{
\item{stats_list}{list of (1) single test stat (R2, evenness, skew) from focal distribution and (2) vector of test stats from sample distributions}
}
\value{
focal_quantile of focal vs. sample
}
\description{
Calculate the quantile of a test stat for a focal RAD vs. a sample of RADs
}
