% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_emp_constraint.R
\name{test_quantile}
\alias{test_quantile}
\title{Calculate the quantile of a test stat for a focal RAD vs. a sample of RADs}
\usage{
test_quantile(focal_stat, sample_stats)
}
\arguments{
\item{focal_stat}{Single test stat (R2, evenness, skew) from focal distribution}

\item{sample_stats}{Vector of test stats from sample distributions}
}
\value{
focal_quantile
}
\description{
Calculate the quantile of a test stat for a focal RAD vs. a sample of RADs
}
